# marketplace-app/src/main/resources/openapi/marketplace-openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 0.1.0
  description: |
    Complete OpenAPI documentation for the Marketplace backend.
    All business endpoints are prefixed with /api/v1.
    Security: endpoints require bearerAuth (JWT) unless noted otherwise.
    Admin endpoints require the ADMIN role.
servers:
  - url: /
paths:
...
  /api/v1/listings:
    get:
      summary: List all listings
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated list of listing summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_ListingSummary'
    post:
      summary: Create a new listing
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '200':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingResponse'
...
  /api/v1/listings/{id}/archive:
    post:
      summary: Archive a listing
      tags: [Catalog]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Listing archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingResponse'

  # ─── Booking Module ─────────────────────────────────────────────
  /api/v1/bookings:
    post:
      summary: Create a booking
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '200':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/v1/bookings/{id}:
    get:
      summary: Get booking by ID
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/v1/bookings/consumer/{consumerId}:
    get:
      summary: List bookings by consumer
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: consumerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated bookings by consumer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_BookingResponse'
  /api/v1/bookings/provider/{providerId}:
    get:
      summary: List bookings by provider
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated bookings by provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_BookingResponse'
  /api/v1/bookings/{id}/confirm:
    post:
      summary: Confirm a booking
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/v1/bookings/{id}/complete:
    post:
      summary: Complete a booking
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/v1/bookings/{id}/cancel:
    post:
      summary: Cancel a booking
      tags: [Booking]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'

  # ─── Payments Module ────────────────────────────────────────────
...
  /api/v1/admin/ledger/providers/{providerId}/balance:
    get:
      summary: Get provider ledger balance (admin)
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Provider balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderBalance'
  /api/v1/admin/ledger/providers/{providerId}/credit:
    post:
      summary: Credit a provider's balance (admin)
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amountCents
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: sourceId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Balance credited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderBalance'

  # ─── Disputes Module ────────────────────────────────────────────
...
  /api/v1/admin/users:
    get:
      summary: List all users (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated user summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_UserSummary'
  /api/v1/admin/listings:
    get:
      summary: List all listings (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated listing summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_ProviderListingSummary'
  /api/v1/admin/listings/{id}/archive:
    post:
      summary: Archive a listing (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Listing archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderListingSummary'
  /api/v1/admin/bookings:
    get:
      summary: List all bookings (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated booking summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_BookingSummary'
  /api/v1/admin/payments:
    get:
      summary: List all payment intents (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated payment summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_PaymentSummary'
  /api/v1/admin/payments/{id}:
    get:
      summary: Get payment intent by ID (admin)
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSummary'
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
      description: Zero-based page index
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Page size
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Sort property, e.g. createdAt,desc
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Provider ──────────────────────────────────────────────────
...
    CreateListingRequest:
      type: object
      required: [title, category, priceCents]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        priceCents:
          type: integer
          format: int64
...
    ListingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListingSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        providerName:
          type: string
    ProviderListingSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        providerId:
          type: string
          format: uuid
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateBookingRequest:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: string
          format: uuid
        notes:
          type: string
    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        consumerId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        status:
          type: string
        priceCents:
          type: integer
          format: int64
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaymentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        consumerId:
          type: string
          format: uuid
        amountCents:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProviderBalance:
      type: object
      properties:
        providerId:
          type: string
          format: uuid
        availableCents:
          type: integer
          format: int64

    # ── Disputes ──────────────────────────────────────────────────
...
    PagedResponse_UserSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    PagedResponse_ListingSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ListingSummary'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
...
    PagedResponse_BookingResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    PagedResponse_BookingSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookingSummary'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    PagedResponse_PaymentSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PaymentSummary'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    PagedResponse_ProviderListingSummary:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProviderListingSummary'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
...